reorg and separation of base models and builders/extensions
