+++ Sources/Models/Extensions/CustomEmoji+Extensions.swift
@@ -1,0 +1,13 @@
+//
+//  CustomEmoji+Extensions.swift
+//  Chameleon
+//
+//  Created by Ian Keen on 15/06/2016.
+//
+//
+
+extension CustomEmoji: SlackEmoji {
+    public var emojiSymbol: String {
+        return ":\(self.name):"
+    }
+}
