+++ Sources/Models/Group.swift
@@ -1,6 +1,6 @@
 //
 //  Group.swift
-// Chameleon
+//  Chameleon
 //
 //  Created by Ian Keen on 25/05/2016.
 //  Copyright © 2016 Mustard. All rights reserved.
@@ -27,26 +27,4 @@ public struct Group {
     
     public let unread_count: Int?
     public let unread_count_display: Int?
-}
-
-extension Group: SlackModelType {
-    public static func make(builder: SlackModelBuilder) throws -> Group {
-        return try tryMake(Group(
-            id:                     try builder.property("id"),
-            name:                   try builder.property("name"),
-            created:                try builder.property("created"),
-            creator:                try builder.slackModel("creator"),
-            is_group:               builder.property("is_group"),
-            is_archived:            builder.property("is_archived"),
-            is_mpim:                builder.property("is_mpim"),
-            members:                try builder.optionalSlackModels("members"),
-            topic:                  try builder.optionalProperty("topic"),
-            purpose:                try builder.optionalProperty("purpose"),
-            last_read:              builder.optionalProperty("last_read"),
-            latest:                 try builder.optionalProperty("latest"),
-            unread_count:           builder.optionalProperty("unread_count"),
-            unread_count_display:   builder.optionalProperty("unread_count_display")
-            )
-        )
-    }
 }
