+++ Sources/Models/Purpose.swift
@@ -11,14 +11,3 @@ public struct Purpose {
     public let creator: User?
     public let last_set: Int
 }
-
-extension Purpose: SlackModelType {
-    public static func make(builder: SlackModelBuilder) throws -> Purpose {
-        return try tryMake(Purpose(
-            value:      try builder.property("value"),
-            creator:    try builder.optionalSlackModel("creator"),
-            last_set:   try builder.property("last_set")
-            )
-        )
-    }
-}
