+++ Sources/Models/Topic.swift
@@ -11,14 +11,3 @@ public struct Topic {
     public let creator: User?
     public let last_set: Int
 }
-
-extension Topic: SlackModelType {
-    public static func make(builder: SlackModelBuilder) throws -> Topic {
-        return try tryMake(Topic(
-            value:      try builder.property("value"),
-            creator:    try builder.optionalSlackModel("creator"),
-            last_set:   try builder.property("last_set")
-            )
-        )
-    }
-}
